Fixes a number of issues with the new tcp/serial code where disconnect() was not called during shutdown, resulting in an unclean shutdown. Renamed and moved a few classes and methods to more closely adhere to the previous code. Changing the communications type now forces a disconnect so that the new communications type can immediately be connected. Removed a few unused attributes from the communications interface. Line ending looks like it was intended to be user exposed, but that doesn't work in interfaces, so need to revisit that as it's a good feature.
