+++ src/main/java/org/openpnp/machine/openbuilds/OpenBuildsDriver.java
@@ -16,7 +16,7 @@ import org.openpnp.machine.reference.ReferenceActuator;
 import org.openpnp.machine.reference.ReferenceHead;
 import org.openpnp.machine.reference.ReferenceHeadMountable;
 import org.openpnp.machine.reference.ReferenceNozzle;
-import org.openpnp.machine.reference.driver.AbstractCommunications;
+import org.openpnp.machine.reference.driver.AbstractReferenceDriver;
 import org.openpnp.model.LengthUnit;
 import org.openpnp.model.Location;
 import org.openpnp.spi.Nozzle;
@@ -25,11 +25,7 @@ import org.openpnp.util.Utils2D;
 import org.pmw.tinylog.Logger;
 import org.simpleframework.xml.Attribute;
 
-public class OpenBuildsDriver extends AbstractCommunications implements Runnable {
-    public void close(){
-
-    }
-
+public class OpenBuildsDriver extends AbstractReferenceDriver implements Runnable {
     @Attribute(required = false)
     protected double feedRateMmPerMinute = 5000;
