+++ src/main/java/org/openpnp/machine/reference/driver/wizards/AbstractReferenceDriverConfigurationWizard.java
@@ -7,7 +7,7 @@ import javax.swing.event.PopupMenuListener;
 
 import org.openpnp.gui.support.AbstractConfigurationWizard;
 import org.openpnp.gui.support.IntegerConverter;
-import org.openpnp.machine.reference.driver.AbstractCommunications;
+import org.openpnp.machine.reference.driver.AbstractReferenceDriver;
 import org.openpnp.machine.reference.driver.SerialPortCommunications;
 
 import com.jgoodies.forms.layout.ColumnSpec;
@@ -20,8 +20,8 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.KeyEvent;
 
-public class AbstractCommunicationsConfigurationWizard extends AbstractConfigurationWizard implements ActionListener {
-    private final AbstractCommunications driver;
+public class AbstractReferenceDriverConfigurationWizard extends AbstractConfigurationWizard implements ActionListener {
+    private final AbstractReferenceDriver driver;
     private JComboBox comboBoxPort;
     private JComboBox comboBoxBaud;
     private JComboBox flowControlComboBox;
@@ -37,7 +37,7 @@ public class AbstractCommunicationsConfigurationWizard extends AbstractConfigura
     private JPanel panelTcp;
     private JTextField commsMethod;
 
-    public AbstractCommunicationsConfigurationWizard(AbstractCommunications driver) {
+    public AbstractReferenceDriverConfigurationWizard(AbstractReferenceDriver driver) {
         this.driver = driver;
 
         contentPanel.setLayout(new BoxLayout(contentPanel, BoxLayout.Y_AXIS));
